addressed code-review minor feedback

+++ Source/SGDLib/SGD.cpp
@@ -9,7 +9,7 @@
 #include "DataReaderHelpers.h"
 #include "MatrixQuantizerImpl.h"
 #ifdef QUANTIZED_GRADIENT_AGGREGATION
-static inline bool operator==(const std::pair<double,size_t>& a, size_t b) { assert(b==0); return a.first == b; }
+static inline bool operator==(const std::pair<double,size_t>& a, double b) { assert(b==0); return a.first == b; }
 // ^^ workaround until this line in AggregateGradientsImpl() gets updated: assert(headerCPU->evalErrors[i] == 0);
 #include "AllReduceDistGradAggregator.h"
 #endif
